\name{fun_dat}
\alias{fun_dat}
\docType{data}
\title{
A dataset of 2726 fungal barcode sequences belonging to 1363 fungal species, where each species has exactly 2 sequence.
}
\description{
These sequences has been collected from BOLD systems.
}
\usage{data (fun_dat)}

\source{
BOLD systems: \url{ http://www.boldsystems.org/ }
}
\references{
\enumerate{
\item Ratnasingham S., and Hebert P.D.N. (2007). BOLD: the barcode of life data system. \emph{Mol. Ecol. Notes}, 7, 355-364.
}
}
\examples{
data (fun_dat)

}
\keyword{datasets}
